Minimal CMake support
This adds a CMakeLists for building using cmake. This means
that this project can be added as a sub module (CPM or a simple subdir)
and then just depend on it.

Code is left unchanged on purpose... even tough some C++ guards should be in added.

To build using CMake run (you can change the build dir as needed):

	cmake -S . -B cbuild
	cmake --build cbuild/